Fix compatibility with iced 0.14 API changes
- Remove Default bound from Application::Theme (iced Theme enum doesn't impl Default)
- Fix viewport scale factor f64→f32 casts
- Rewrite clipboard for new Content/Error API
- Add keyboard event repeat field, update mouse interaction enum mapping
- Update keyboard module re-exports

+++ src/application/state.rs
@@ -119,7 +119,7 @@ where
                         window_info.physical_size().width,
                         window_info.physical_size().height,
                     ),
-                    scale,
+                    scale as f32,
                 );
 
                 self.viewport_version = self.viewport_version.wrapping_add(1);
@@ -183,7 +183,7 @@ where
 
                     self.viewport = Viewport::with_physical_size(
                         self.viewport.physical_size(),
-                        self.system_scale_factor,
+                        self.system_scale_factor as f32,
                     );
 
                     self.viewport_version = self.viewport_version.wrapping_add(1);
@@ -201,7 +201,7 @@ where
                     self.scale_policy = WindowScalePolicy::ScaleFactor(*new_scale);
 
                     self.viewport =
-                        Viewport::with_physical_size(self.viewport.physical_size(), *new_scale);
+                        Viewport::with_physical_size(self.viewport.physical_size(), *new_scale as f32);
 
                     self.viewport_version = self.viewport_version.wrapping_add(1);
                 }
